Prying apart the fs module for maintainability
