+++ src/fs/cluster.rs
@@ -1,5 +1,8 @@
+use crate::sector::Sector;
+
+#[derive(Debug, PartialEq)]
 pub struct Cluster {
+    sectors: Vec<Sector>,
 }
 
-impl Cluster {
-}
+impl Cluster {}
